renamed to BookDetailsParserFromHtml based on working
